0.27.0
Added basic redirection logic to dashboard home page to verify a user is loggin before hitting the dash otherwise redirect to login page, updated components/home/pricingcards.tsx/ to dynamically render over pkg data from the db via the gql server using a newly added query method. Updated the logic so that and id is retrived from the event parent to be used in the mutation that creates the session so that price is reflected depending on which pkg card is clicked. added new query to lib/fragments/queries.ts to get all pkgs (price plans).